# backend/app/routes/analytics.py
from __future__ import annotations

import httpx
from fastapi import APIRouter, Depends, HTTPException

from app.deps.github_client import get_github_client
from app.schemas.analytics import RepoAnalyticsResponse
from app.schemas.compare import CompareReposRequest, CompareReposResponse
from app.schemas.insights import InsightsEnhanceRequest, InsightsSummaryResponse
from app.schemas.prediction import ContributionPredictionResponse
from app.services.contribution_prediction import predict_contribution_trend
from app.services.github import GitHubApiError, GitHubClient
from app.services.insights import build_insights_summary
from app.services.repo_compare import compare_repositories
from app.services.repository_analytics import get_repo_analytics

router = APIRouter(prefix="/analytics", tags=["analytics"])


@router.get("/repo/{owner}/{repo}", response_model=RepoAnalyticsResponse)
async def repo_analytics(owner: str, repo: str) -> RepoAnalyticsResponse:
    """
    Returns analytics for a repository based on stored snapshots.

    Reads the two most recent snapshots from the database and computes
    the difference in stars and forks between them (stars_delta, forks_delta).

    Note: This endpoint does NOT call the GitHub API — it only reads from
    the database. To get fresh data, hit GET /api/repo/{owner}/{repo} first,
    which saves a new snapshot.

    Raises:
    - 404 if no snapshots exist yet for this repo (i.e. it has never been fetched)
    """
    analytics = await get_repo_analytics(owner, repo)
    if analytics is None:
        raise HTTPException(status_code=404, detail="No stored repository snapshots for this repo yet.")
    return analytics


@router.get("/repo/{owner}/{repo}/insights", response_model=InsightsSummaryResponse)
async def repo_insights(
    owner: str,
    repo: str,
    gh: GitHubClient = Depends(get_github_client),
) -> InsightsSummaryResponse:
    """
    Health score, risk list, and heuristic high-impact insight lines.

    Requires prior snapshots so `stars_delta` is defined; client should call
    GET /api/repo/{owner}/{repo} first to persist snapshots.
    """
    analytics = await get_repo_analytics(owner, repo)
    if analytics is None:
        raise HTTPException(
            status_code=404,
            detail="No stored repository snapshots for this repo yet. Fetch /api/repo/{owner}/{repo} first.",
        )

    try:
        return await build_insights_summary(
            owner,
            repo,
            github_client=gh,
            stars_delta=analytics.stars_delta,
            use_openai=False,
        )
    except GitHubApiError as e:
        raise HTTPException(status_code=e.status_code or 502, detail=e.message) from e
    except httpx.HTTPError as e:
        raise HTTPException(status_code=502, detail=str(e)) from e


@router.post("/repo/{owner}/{repo}/insights", response_model=InsightsSummaryResponse)
async def repo_insights_with_options(
    owner: str,
    repo: str,
    body: InsightsEnhanceRequest,
    gh: GitHubClient = Depends(get_github_client),
) -> InsightsSummaryResponse:
    """
    Same metrics as GET, but when `use_openai` is true and an API key is available
    (request body or server OPENAI_API_KEY), high_impact_insights are generated by the model
    instead of heuristics.
    """
    analytics = await get_repo_analytics(owner, repo)
    if analytics is None:
        raise HTTPException(
            status_code=404,
            detail="No stored repository snapshots for this repo yet. Fetch /api/repo/{owner}/{repo} first.",
        )

    try:
        return await build_insights_summary(
            owner,
            repo,
            github_client=gh,
            stars_delta=analytics.stars_delta,
            use_openai=body.use_openai,
            openai_api_key=body.openai_api_key,
        )
    except GitHubApiError as e:
        raise HTTPException(status_code=e.status_code or 502, detail=e.message) from e
    except httpx.HTTPError as e:
        raise HTTPException(status_code=502, detail=str(e)) from e


@router.get("/repo/{owner}/{repo}/prediction", response_model=ContributionPredictionResponse)
async def repo_contribution_prediction(
    owner: str,
    repo: str,
    gh: GitHubClient = Depends(get_github_client),
) -> ContributionPredictionResponse:
    """Weekly-commit-based trend and rough next-week estimate (no DB snapshots required)."""
    try:
        return await predict_contribution_trend(owner, repo, github_client=gh)
    except GitHubApiError as e:
        raise HTTPException(status_code=e.status_code or 502, detail=e.message) from e
    except httpx.HTTPError as e:
        raise HTTPException(status_code=502, detail=str(e)) from e


@router.post("/compare/repos", response_model=CompareReposResponse)
async def compare_repos(
    body: CompareReposRequest,
    gh: GitHubClient = Depends(get_github_client),
) -> CompareReposResponse:
    """Body: 2–3 `{owner, repo}` entries; response rows align with request order."""
    try:
        return await compare_repositories(body.repos, github_client=gh)
    except GitHubApiError as e:
        raise HTTPException(status_code=e.status_code or 502, detail=e.message) from e
    except httpx.HTTPError as e:
        raise HTTPException(status_code=502, detail=str(e)) from e