# backend/openapi.yaml
# =============================================================================
# FRONTEND / LEAD CONTRACT — do not rename or remove paths without team agreement.
# -----------------------------------------------------------------------------
# The FastAPI app serves this file verbatim as GET /openapi.json (see app/openapi_spec.py).
# Backend implementations must match these paths and methods.
#
# CHANGE LOG (for co-developers):
# - 2026-03: Documented how each path maps to GitHub REST (backend proxies GitHub JSON).
# - GET /api/repo/{owner}/{repo}/issue-stats returns a small aggregated object
#   { open_issues, closed_issues, total_issues } via GitHub Search API (not a single
#   upstream repo endpoint). Frontend should treat this shape as stable.
# - GET /api/repo/{owner}/{repo}/activity maps to GitHub
#   GET /repos/{owner}/{repo}/stats/commit_activity; GitHub may respond 202 while
#   stats build — backend responds with 503 and a clear message in that case.
# - GET /api/repo/{owner}/{repo}/tree resolves default_branch then returns
#   GET /repos/{owner}/{repo}/git/trees/{sha}?recursive=1.
# - Optional query param per_page (1–100, default 30) is accepted on list endpoints
#   where GitHub supports it; it was not in the original spec — safe for clients to omit.
# - POST /api/repo/{owner}/{repo}/generate-readme returns structured README markdown/HTML
#   from GitHub data, snapshot deltas when present, and optional OpenAI narrative.
#
# NOT IN THIS FILE (ops / future — not part of the frontend contract until added here):
# - GET /health — process health only.
# - Experimental routes may exist in app/routes/ (e.g. analytics, OAuth, favorites) but are
#   not mounted in main.py until this spec is updated by the lead.
# =============================================================================

openapi: 3.0.3
info:
  title: GitHub Project Visualizer API
  version: 1.0.0

servers:
  - url: http://localhost:8000

paths:
  # Proxies GitHub GET /repos/{owner}/{repo}; response normalized to app RepoResponse for this path only.
  /api/repo/{owner}/{repo}:
    get:
      summary: Get repository overview
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repository info
        "404":
          description: Repo not found

  # Proxies GitHub GET /repos/{owner}/{repo}/commits
  /api/repo/{owner}/{repo}/commits:
    get:
      summary: Get commit history
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of commits
        "404":
          description: Repo not found

  # Proxies GitHub GET /repos/{owner}/{repo}/contributors
  /api/repo/{owner}/{repo}/contributors:
    get:
      summary: Get repository contributors
      description: Proxies GitHub contributors list with per_page (1–100). Response fields are a stable subset of the upstream user objects.
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        "200":
          description: Contributors ordered by commit count (descending)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [login, contributions]
                  properties:
                    login:
                      type: string
                    id:
                      type: integer
                    avatar_url:
                      type: string
                      nullable: true
                    html_url:
                      type: string
                      nullable: true
                    contributions:
                      type: integer
                      minimum: 0
        "404":
          description: Repo not found
        "502":
          description: Unexpected upstream response or parse error

  # Proxies GitHub GET /repos/{owner}/{repo}/languages
  /api/repo/{owner}/{repo}/languages:
    get:
      summary: Get language breakdown
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Language percentages
        "404":
          description: Repo not found

  # Proxies GitHub GET /repos/{owner}/{repo}/issues
  /api/repo/{owner}/{repo}/issues:
    get:
      summary: Get issues
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed, all]
      responses:
        "200":
          description: List of issues
        "404":
          description: Repo not found

  # Proxies GitHub GET /repos/{owner}/{repo}/pulls
  /api/repo/{owner}/{repo}/pulls:
    get:
      summary: Get pull requests
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed, all]
      responses:
        "200":
          description: List of pull requests
        "404":
          description: Repo not found

  # Proxies GitHub GET /repos/{owner}/{repo}/branches
  /api/repo/{owner}/{repo}/branches:
    get:
      summary: Get all branches
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of branches
        "404":
          description: Repo not found

  # Proxies GitHub GET /repos/{owner}/{repo}/tags (tags only; releases are separate upstream)
  /api/repo/{owner}/{repo}/tags:
    get:
      summary: Get all tags and releases
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tags
        "404":
          description: Repo not found

  # Proxies GitHub GET /repos/{owner}/{repo}/stats/commit_activity (see header note on 202 → 503)
  /api/repo/{owner}/{repo}/activity:
    get:
      summary: Get commit activity over time
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Weekly commit frequency data
        "404":
          description: Repo not found

  # Aggregated via GitHub Search API — response shape { open_issues, closed_issues, total_issues }
  /api/repo/{owner}/{repo}/issue-stats:
    get:
      summary: Get aggregated issue statistics
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Open vs closed counts, label distribution
        "404":
          description: Repo not found

  # default_branch → git/trees/{sha}?recursive=1 (see header)
  /api/repo/{owner}/{repo}/tree:
    get:
      summary: Get repository file and folder tree
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Directory and file structure
        "404":
          description: Repo not found

  # Proxies GitHub GET /repos/{owner}/{repo}/stargazers
  /api/repo/{owner}/{repo}/stargazers:
    get:
      summary: Get users who starred the repo
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of stargazers
        "404":
          description: Repo not found

  /api/repo/{owner}/{repo}/generate-readme:
    post:
      summary: Generate structured README from repository metrics
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                template_id:
                  type: string
                  default: default
                export_format:
                  type: string
                  enum: [markdown, html, both]
                  default: both
                use_openai:
                  type: boolean
                  default: false
                openai_api_key:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Markdown body and optional HTML export
        "404":
          description: Repo not found
        "502":
          description: GitHub API or auxiliary service error

  # Proxies GitHub GET /rate_limit (authenticated as server PAT or user token when wired)
  /api/rate-limit:
    get:
      summary: Get current GitHub API rate limit status
      responses:
        "200":
          description: Remaining requests and reset time
        "401":
          description: Unauthorized

  /api/analytics/repo/{owner}/{repo}:
    get:
      summary: Get repository analytics from stored snapshots
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Latest and previous snapshot with stars/forks deltas
        "404":
          description: No snapshots found for this repo

  /api/analytics/repo/{owner}/{repo}/insights:
    get:
      summary: Get repository health score, risk signals, and heuristic high-impact insight lines
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Health score, activity trend, risk signals, high_impact_insights, used_openai
        "404":
          description: No snapshots found for this repo
        "502":
          description: GitHub API error
    post:
      summary: Same as GET with optional OpenAI-generated insight bullets
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                use_openai:
                  type: boolean
                  default: false
                openai_api_key:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Insights payload; used_openai true when the model returned bullets
        "404":
          description: No snapshots found for this repo
        "502":
          description: GitHub or OpenAI error

  /api/analytics/repo/{owner}/{repo}/prediction:
    get:
      summary: Heuristic next-week commit estimate from weekly commit totals
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trend direction and projected weekly commits
        "502":
          description: GitHub API error

  /api/analytics/compare/repos:
    post:
      summary: Compare two or three repositories side by side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repos]
              properties:
                repos:
                  type: array
                  minItems: 2
                  maxItems: 3
                  items:
                    type: object
                    required: [owner, repo]
                    properties:
                      owner:
                        type: string
                      repo:
                        type: string
      responses:
        "200":
          description: Comparison rows with health, merge rate, and snapshot deltas
        "502":
          description: GitHub API error